Update to explicitly use the CNCF CoC
It's an oversight that our CoC wasn't the same as the CNCF's. In
practice we've been using the CNCF CoC for a long time - it's the CNCF
CoC which we refer to in meetings.

This PR explicitly uses the CNCF CoC, which is a requirement for
Graduation.

See also the Slack announcement of this change:

https://kubernetes.slack.com/archives/CDEQJ0Q8M/p1721899040392779

Signed-off-by: Ashley Davis <ashley.davis@venafi.com>